% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplotWrapper.R
\name{flexplot}
\alias{flexplot}
\title{Flexplot}
\usage{
flexplot(
  data = NULL,
  version = "1",
  formula = NULL,
  alpha = 0.4,
  bw = FALSE,
  confidence = FALSE,
  dependent = list(types = list(), value = ""),
  ghost = TRUE,
  intervals = "Quartiles",
  jitx = 0.1,
  jity = 0,
  nameCols = "",
  nameLegend = "",
  nameRows = "",
  nameX = "",
  nameY = "",
  palette = "GGplot Default",
  paneledVars = list(types = list(), value = list()),
  plotHeight = 320,
  plotWidth = 480,
  theme = "JASP",
  type = "Loess",
  variables = list(types = list(), value = list())
)
}
\arguments{
\item{alpha, }{The degree of transparency of the dots in the graphics}

\item{confidence, }{Should 95% confidence intervals be displayed?
Defaults to \code{FALSE}.}

\item{dependent, }{The variable of interest. This is also called the outcome variable. If only an outcome variable is specified, Flexplot will produce a histogram for numeric data and a barchart for categorical data. If independent variable(s) and/or panelled variable(s) are specified, this variable will be displayed on the Y axis.}

\item{ghost, }{Ghost lines are a visual aid that can be used when doing panelled plots. Ghost lines simply repeat the fitted line from one panel across the other panels to make it easier to make comparisons across panels.
Defaults to \code{TRUE}.}

\item{jitx, }{The maximal amount of "jittering" used on the X axis to remove overlap between datapoints. The maximal amount of jittering is proportional to the density of the data. In other words, the maximum jittering will occur at the mode of the dataset, and little to no jittering will occur in locations where there is no overlap.}

\item{jity, }{The maximal amount of "jittering" used on the Y axis to remove overlap between datapoints. The maximal amount of jittering is proportional to the density of the data. In other words, the maximum jittering will occur at the mode of the dataset, and little to no jittering will occur in locations where there is no overlap.}

\item{paneledVars, }{Variables specified in these boxes will be binned (if numeric) then displayed as different subplots. The first variable specified as a panelled variable will form the column plots, while the second will form the row plots.}

\item{theme, }{the type of GGplot theme to use when displaying the data. Can be one of the following:}

\item{type, }{The type of fitted line displayed when the x axis is a numeric variable. These can be one of the following:
\itemize{
  \item \code{"Loess"} (default) : A non-parametric loess line
  \item \code{"Regression"}: A straight (regression) line
  \item \code{"Quadratic"}: A line that includes both a linear effect and a quadratic (squared) term
  \item \code{"Cubic"}: A line that includes a linear, squared, and cubed term
  \item \code{"None"}: No line
}}

\item{variables, }{The variable(s) for which we wish to visually assess relationships with the DV. The first variable chosen shows up on the X axis, either as a scatterplot (for numeric predictors) or as a beeswarm plot (for categorical variables). The second variable chosen will show up as different colors/lines/symbols. If the second varaible chosen is numeric, it will be binned first.}
}
\description{
Flexplot allows the user to create graphical displays of data, using barcharts and histograms for univariate data, and various different types of scatterplots for bivariate and multivariate data.
}
