% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspPlotBuilderWrapper.R
\name{jaspPlotBuilder}
\alias{jaspPlotBuilder}
\title{Plot Builder}
\usage{
jaspPlotBuilder(
  data = NULL,
  version = "1",
  PlotBuilderTab = list(list(XPaddingFirst = 0.05, XPaddingSecond = 0.05, YPaddingFirst =
    0.04, YPaddingSecond = 0.05, accuracyCountValue = 0.1, accuracyMeanValue = 0.1,
    accuracyMedianValue = 0.1, accuracySumValue = 0.1, addAreaStack = FALSE, addBarStack
    = FALSE, addBoxplot = FALSE, addCI95ErrorBar = FALSE, addCi95Ribbon = FALSE,
    addCountArea = FALSE, addCountBar = FALSE, addCountDash = FALSE, addCountDot = FALSE,
    addCountLine = FALSE, addCountValue = FALSE, addCurveFitPlotBuilder = FALSE,
    addDataPoint = FALSE, addDensity = FALSE, 
     addHistogram = FALSE,
    addIdentityLinePlotBuilder = FALSE, addMeanArea = FALSE, addMeanBar = FALSE,
    addMeanDash = FALSE, addMeanDot = FALSE, addMeanLine = FALSE, addMeanValue = FALSE,
    addMedianArea = FALSE, addMedianBar = FALSE, addMedianDash = FALSE, addMedianDot =
    FALSE, addMedianLine = FALSE, addMedianValue = FALSE, addRangeErrorBar = FALSE,
    addRangeRibbon = FALSE, addReferenceLinePlotBuilder = FALSE, addSDErrorBar = FALSE,
    addSEMErrorBar = FALSE, addSdRibbon = FALSE, addSemRibbon = FALSE, 
    
    addStatEllipse = FALSE, addSumArea = FALSE, addSumBar = FALSE, addSumDash = FALSE,
    addSumDot = FALSE, addSumLine = FALSE, addSumValue = FALSE, addViolin = FALSE,
    aggregationFun = "mean", aggregationFunY = "mean", alphaAreaStack = 0.4,
    alphaBarStack = 0.8, alphaBoxplotPlotBuilder = 0.8, alphaCi95Ribbon = 0.4,
    alphaCountArea = 1, alphaCountBar = 1, alphaCountDash = 1, alphaCountDot = 1,
    alphaCountLine = 1, alphaCountValue = 1, alphaDensityPlotBuilder = 0.8,
    alphaHistogramPlotBuilder = 0.8, alphaMeanArea = 1, 
     alphaMeanBar = 1,
    alphaMeanDash = 1, alphaMeanDot = 1, alphaMeanLine = 1, alphaMeanValue = 1,
    alphaMedianArea = 1, alphaMedianBar = 1, alphaMedianDash = 1, alphaMedianDot = 1,
    alphaMedianLine = 1, alphaMedianValue = 1, alphaPlotBuilder = 0.5, alphaRangeRibbon =
    0.4, alphaSdRibbon = 0.4, alphaSemRibbon = 0.4, alphaSumArea = 1, alphaSumBar = 1,
    alphaSumDash = 1, alphaSumDot = 1, alphaSumLine = 1, alphaSumValue = 1,
    alphaViolinPlotBuilder = 0.8, annotationLineList = list(), annotationPlotBuilder =
    list(), 
     asPercentage = FALSE, asTable = "bottom-right", asTableFacetWrap =
    "bottom-rightFacetWrap", axes = "margins", axisLabels = "all", backgroundDataLabels =
    FALSE, barwidthCountBar = 0.8, baseFontSize = 18, betweenSubjectFactors = list(types
    = list(), value = list()), binsPlotBuilder = 30, blackDensityOutline = FALSE,
    blackHistogramOutline = FALSE, blackOutlineBoxplot = TRUE, blackOutlineCI95ErrorBar =
    FALSE, blackOutlineCi95Ribbon = FALSE, blackOutlineCountDash = FALSE,
    blackOutlineCountDot = FALSE, 
     blackOutlineCountLine = FALSE,
    blackOutlineCountValue = FALSE, blackOutlineCurveFit = FALSE, blackOutlineDataPoint =
    FALSE, blackOutlineMeanDash = FALSE, blackOutlineMeanDot = FALSE,
    blackOutlineMeanLine = FALSE, blackOutlineMeanValue = FALSE, blackOutlineMedianDash =
    FALSE, blackOutlineMedianDot = FALSE, blackOutlineMedianLine = FALSE,
    blackOutlineMedianValue = FALSE, blackOutlineRangeErrorBar = FALSE,
    blackOutlineRangeRibbon = FALSE, blackOutlineSDErrorBar = FALSE,
    blackOutlineSEMErrorBar = FALSE, 
     blackOutlineSdRibbon = FALSE,
    blackOutlineSemRibbon = FALSE, blackOutlineSumDash = FALSE, blackOutlineSumDot =
    FALSE, blackOutlineSumLine = FALSE, blackOutlineSumValue = FALSE, blackOutlineViolin
    = FALSE, bottomMargin = 10, breakByX = "", breakByY = "", breakFromX = "", breakFromY
    = "", breakToX = "", breakToY = "", captionPlotBuilder = "", colSplitRM = list(types
    = list(), value = ""), colorByGroup = "none", colorIdentityLine = "lightgray",
    colorLabelRenamer = list(), colorReferenceLine = "lightgray", 
     colorsAll =
    "jaspPalette", columnsvariableSplitPlotBuilder = list(types = list(), value = ""),
    connectRMPlotBuilder = FALSE, covariates = list(types = list(), value = list()),
    curvaFitMethod = "lm", customColors = "", cutShortScale = FALSE, cutShortScaleY =
    FALSE, dashwidthCountDash = 0.8, dashwidthMeanDash = 0.8, dashwidthMedianDash = 0.8,
    deleteNAListwise = TRUE, densityOverlayMode = "count", dodgeBoxplotPlotBuilder = 0.8,
    dodgeCI95ErrorBar = 0.8, dodgeCountArea = 0.8, dodgeCountBar = 0.8, 
    
    dodgeCountDash = 0.8, dodgeCountDot = 0.8, dodgeCountLine = 0.8, dodgeCurveFit = 0.8,
    dodgeMeanArea = 0.8, dodgeMeanBar = 0.8, dodgeMeanDash = 0.8, dodgeMeanDot = 0.8,
    dodgeMeanLine = 0.8, dodgeMedianArea = 0.8, dodgeMedianBar = 0.8, dodgeMedianDash =
    0.8, dodgeMedianDot = 0.8, dodgeMedianLine = 0.8, dodgeRangeErrorBar = 0.8,
    dodgeSDErrorBar = 0.8, dodgeSEMErrorBar = 0.8, dodgeSumArea = 0.8, dodgeSumBar = 0.8,
    dodgeSumDash = 0.8, dodgeSumDot = 0.8, dodgeSumLine = 0.8, dodgeViolinPlotBuilder =
    0.8, 
     drawQuantilesViolinPlotBuilder = "0.25, 0.5, 0.75", ellipseType = "t",
    emptyCircles = FALSE, enableSort = FALSE, enableSortY = FALSE, fillEllipse = 0.2,
    flipPlot = FALSE, fontsizeCountValue = 14, fontsizeDataLabels = 4, fontsizeMeanValue
    = 14, fontsizeMedianValue = 14, fontsizeSumValue = 14, gridVarRM = list(types =
    list(), value = ""), gridVariablePlotBuilder = list(types = list(), value = ""),
    groupVarRM = list(types = list(), value = ""), heightPlotBuilder = 300,
    hjustCountValue = 0.5, hjustMeanValue = 0.5, 
     hjustMedianValue = 0.5,
    hjustSumValue = 0.5, isRM = "noRM", jitterhPlotBuilder = 0.3, jitterwPlotBuilder =
    0.3, labelSizePValue = 4.5, labelVariablePlotBuilder = list(types = list(), value =
    ""), labelcolor = "black", layerOrder = "", leftMargin = 10,
    legendPosistionPlotBuilder = "right", levelEllipse = 0.95, limitFromX = "",
    limitFromY = "", limitToX = "", limitToY = "", lineRMsize = 0.5, lineRMtransparency =
    0.5, lineWidthDensity = 0.8, linewidhtReferenceLines = 1, linewidthAreaStack = 0.25,
    
     linewidthCI95ErrorBar = 1, linewidthCountDash = 1, linewidthCountLine = 1,
    linewidthCurveFit = 1, linewidthMeanDash = 1, linewidthMeanLine = 1,
    linewidthMedianDash = 1, linewidthMedianLine = 1, linewidthRangeErrorBar = 1,
    linewidthSDErrorBar = 1, linewidthSEMErrorBar = 1, linewidthSumDash = 1,
    linewidthSumLine = 0.5, linewidthViolinPlotBuilder = 0.8, margins = FALSE,
    ncolFacetWrap = 1, nrowFacetWrap = 1, outlierBoxplotPlotBuilder = FALSE,
    outlierCoefBoxplotPlotBuilder = 1.5, outlierSizeBoxplotPlotBuilder = 1, 
    
    pairwiseComparisons = list(), plotStyle = "JASP", pointDodgePlotBuilder = 0.8,
    pointShapeVariable = list(types = list(), value = ""), pointSizeMax = 4, pointSizeMin
    = 1, pointsizePlotBuilder = 3, propMode = "absolute", removeLegendTitle = FALSE,
    repeatedMeasuresCells = list("", ""), repeatedMeasuresFactors = list(list(levels =
    list("Level 1", "Level 2"), name = "RM Factor 1")), replaceNaAreaStack = FALSE,
    reverseAreaStack = FALSE, reverseBarStack = FALSE, reversedirectionIdentityLine =
    FALSE, 
     rightMargin = 10, rotateXLabel = FALSE, rotateYLabel = FALSE, rowSplitRM
    = list(types = list(), value = ""), rowsvariableSplitPlotBuilder = list(types =
    list(), value = ""), scaleViolinPlotBuilder = "width", scales = "fixed",
    scalesFacetWrap = "fixed", seCurveFit = TRUE, shapebyVariable = FALSE, sizeByVariable
    = FALSE, sizeCountDot = 5, sizeMeanDot = 5, sizeMedianDot = 5, sizeSumDot = 5,
    sizeVariablePlotBuilder = list(types = list(), value = ""), sortXLabelsOrder =
    "Increasing", sortYLabelsOrder = "Increasing", 
     space = "fixed", stepDistance =
    0.15, stripPosition = "top", titlePlotBuilder = "", titleXPlotBuilder = "",
    titleYPlotBuilder = "", topMargin = 10, transparencyCI95ErrorBar = 1,
    transparencyCurveFit = 0.2, transparencyRangeErrorBar = 1, transparencySDErrorBar =
    1, transparencySEMErrorBar = 1, trimViolinPlotBuilder = FALSE, value = "Plot 1",
    variableColorPlotBuilder = list(types = list(), value = ""), variableXPlotBuilder =
    list(types = list(), value = ""), variableYPlotBuilder = list(types = list(), 
      
      value = ""), vjustCountValue = -0.5, vjustMeanValue = -0.5, vjustMedianValue =
    -0.5, vjustSumValue = -0.5, whiteBorder = FALSE, whiteOutlineBoxplot = FALSE,
    whiteOutlineCountDot = FALSE, whiteOutlineMeanDot = FALSE, whiteOutlineMedianDot =
    FALSE, whiteOutlineSumDot = FALSE, whiteOutlineViolin = FALSE,
    widthBoxplotPlotBuilder = 0.6, widthCI95ErrorBar = 0.3, widthLineBoxplotPlotBuilder =
    0.8, widthMeanBar = 0.8, widthMedianBar = 0.8, widthPlotBuilder = 380,
    widthRangeErrorBar = 0.3, widthSDErrorBar = 0.3, 
     widthSEMErrorBar = 0.3,
    widthSumBar = 0.8, widthSumDash = 0.8, widthWhiskersPlotBuilder = 0.3,
    xAxisLabelRenamer = list(), xAxisTitleSplit = "", xReferenceLine = "", xVarRM =
    list(types = list(), value = ""), yAxisLabelRenamer = list(), yAxisTitleSplit = "",
    yPositionPValue = 70, yReferenceLine = "")),
  columnWidthInput = "",
  fullRowSpecifications = list(),
  getCommonLegend = FALSE,
  labelDistance1 = 0.05,
  labelDistance2 = 0.95,
  labelSize = 18,
  layoutHeight = 500,
  layoutWidth = 500,
  plotHeight = 320,
  plotSpacing = 10,
  plotWidth = 480,
  relHeightWithinRowLayout = "",
  relativeHeight = "",
  rowSpecifications = list()
)
}
\arguments{
\item{columnWidthInput, }{Comma-separated numbers that set the width of each column}

\item{getCommonLegend, }{Place a single shared legend for all plots in the layout
Defaults to \code{FALSE}.}

\item{labelDistance1, }{Horizontal offset of labels (0–1, relative to plot width)}

\item{labelDistance2, }{Vertical offset of labels (0–1, relative to plot height)}

\item{labelSize, }{Font size of subplot labels}

\item{layoutHeight, }{Overall height of the assembled layout (pixels)}

\item{layoutWidth, }{Overall width of the assembled layout (pixels)}

\item{plotSpacing, }{Gap between plots inside the layout (pixels)}

\item{relHeightWithinRowLayout, }{Comma-separated numbers that set the height of each row}

\item{relativeHeight, }{Relative heights of columns or widths of rows when combining column and row layouts}
}
\description{
The <b>Plot Builder (beta)</b> lets you assemble a complete figure in just a few clicks.<br/><br/><ol><li><b>Decide on design:</b> At the very top choose <i>Repeated measurements → Yes</i> if the same subjects are measured more than once (e.g., across time-points or conditions); otherwise leave <i>No</i> selected. </li><li><b>Map your variables:</b> Drag variables from the list on the left into the fields on the right (X-Axis, Y-Axis, etc). </li><li><b>Choose what to display:</b> Expand the <i>Data and geometries</i> accordion to add layers such as raw points, histograms/box-/violin plots, counts, proportions, Mean or Median summaries, etc.</li><li><b>Polish the figure:</b> Use the remaining accordions to fine-tune axes, titles, annotations, themes, colors, sizing, and legend placement.</li>
}
