% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivesWrapper.R
\name{Descriptives}
\alias{Descriptives}
\title{Descriptives}
\usage{
Descriptives(
  data = NULL,
  version = "0.17.1",
  formula = NULL,
  boxPlot = FALSE,
  boxPlotBoxPlot = TRUE,
  boxPlotColourPalette = FALSE,
  boxPlotJitter = FALSE,
  boxPlotOutlierLabel = FALSE,
  boxPlotViolin = FALSE,
  chartType = "_1noCharts",
  chartValues = "_1frequencies",
  ciBootstrapSamples = 1000,
  coefficientOfVariation = FALSE,
  colorPalette = "colorblind",
  correlationPlots = FALSE,
  densityPlot = FALSE,
  densityPlotSeparate = "",
  densityPlotTransparency = 20,
  descriptivesTableTransposed = FALSE,
  distributionAndCorrelationPlotDensity = FALSE,
  distributionAndCorrelationPlotHistogramBinWidthType = "sturges",
  distributionAndCorrelationPlotHistogramManualNumberOfBins = 30,
  distributionAndCorrelationPlotRugMarks = FALSE,
  distributionPlots = FALSE,
  dotPlot = FALSE,
  frequencyTables = FALSE,
  frequencyTablesMaximumDistinctValues = 10,
  heatmapDisplayValue = FALSE,
  heatmapDisplayValueRelativeTextSize = 1,
  heatmapHorizontalAxis = "",
  heatmapLegend = FALSE,
  heatmapStatisticContinuous = "mean",
  heatmapStatisticDiscrete = "mode",
  heatmapTileWidthHeightRatio = 1,
  heatmapVerticalAxis = "",
  intervalPlot = FALSE,
  iqr = FALSE,
  kurtosis = FALSE,
  likertPlot = FALSE,
  likertPlotAdjustableFontSize = "normal",
  likertPlotAssumeVariablesSameLevel = FALSE,
  mad = FALSE,
  madRobust = FALSE,
  maximum = TRUE,
  mean = TRUE,
  meanCi = FALSE,
  meanCiLevel = 0.95,
  meanCiMethod = "normalModel",
  median = FALSE,
  minimum = TRUE,
  missing = TRUE,
  mode = FALSE,
  paretoPlot = FALSE,
  paretoPlotRule = FALSE,
  paretoPlotRuleCi = 0.95,
  percentileValues = list(),
  percentiles = FALSE,
  pieChart = FALSE,
  plotHeight = 320,
  plotWidth = 480,
  qqPlot = FALSE,
  quantilesForEqualGroups = FALSE,
  quantilesForEqualGroupsNumber = 4,
  quartiles = FALSE,
  range = FALSE,
  scatterPlot = FALSE,
  scatterPlotGraphTypeAbove = "density",
  scatterPlotGraphTypeRight = "density",
  scatterPlotLegend = TRUE,
  scatterPlotRegressionLine = TRUE,
  scatterPlotRegressionLineCi = TRUE,
  scatterPlotRegressionLineCiLevel = 0.95,
  scatterPlotRegressionLineType = "smooth",
  sd = TRUE,
  sdCi = FALSE,
  sdCiLevel = 0.95,
  seMean = FALSE,
  shapiroWilkTest = FALSE,
  skewness = FALSE,
  splitBy = "",
  statisticsValuesAreGroupMidpoints = FALSE,
  stemAndLeaf = FALSE,
  stemAndLeafScale = 1,
  sum = FALSE,
  valid = TRUE,
  variables = list(),
  variance = FALSE,
  varianceCi = FALSE,
  varianceCiLevel = 0.95
)
}
\arguments{
\item{correlationPlots}{Displays a matrix of plots between continuous variables, with scatterplots between the variables in the off-diagonal entries, and histograms and density plots in the diagonal entries. The line represents the fit of a 1st, 2nd, 3rd, or 4th order polynomial \insertCite{@the selection is based on the Bayesian information criterion, @schwarz1978estimating}{jaspDescriptives}.}
}
\description{
Descriptives allows the user to obtain basic descriptive statistics, histograms and density plots, correlation plots, boxplots, and frequency tables. For introduction to descriptive analyses please see \insertCite{moore2009introduction,whitlock2015analysis;textual}{jaspDescriptives}.
}
\references{
\insertAllCited{}

\insertRef{ggplot2}{jaspDescriptives}

\insertRef{ggrepel}{jaspDescriptives}
}
