% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptivesTimeSeriesWrapper.R
\name{DescriptivesTimeSeries}
\alias{DescriptivesTimeSeries}
\title{Time Series Descriptives}
\usage{
DescriptivesTimeSeries(
  data = NULL,
  version = "0.95",
  acf = FALSE,
  acfCi = TRUE,
  acfCiLevel = 0.95,
  acfCiType = "whiteNoise",
  acfMaxLag = 10,
  acfZeroLag = FALSE,
  dateEnd = "",
  dateStart = "",
  dependent = list(types = list(), value = ""),
  descriptivesTableTransposed = FALSE,
  filter = FALSE,
  filterBy = "row",
  lagPlot = FALSE,
  lagPlotLag = 1,
  lagPlotRegressionCi = TRUE,
  lagPlotRegressionCiLevel = 0.95,
  lagPlotRegressionLine = TRUE,
  lagPlotRegressionType = "smooth",
  pacf = FALSE,
  pacfCi = TRUE,
  pacfCiLevel = 0.95,
  pacfMaxLag = 10,
  plotHeight = 320,
  plotWidth = 480,
  rowEnd = 100,
  rowStart = 1,
  time = list(types = list(), value = ""),
  timeEnd = 100,
  timeSeriesPlot = TRUE,
  timeSeriesPlotDistribution = "none",
  timeSeriesPlotType = "both",
  timeStart = 1
)
}
\arguments{
\item{acf, }{Plots the autocorrelation for a specified number of lags. The confidence interval may be given assuming either a white noise process, or assuming for a lag q a moving average process of order q - 1.
Defaults to \code{FALSE}.}

\item{dependent, }{A variable that is measured repeatedly over time.}

\item{filter, }{Filters the time series so only a specific range will be used for further analyses. Row number refers to the row number in the spreadsheet. If a 'Time' variable is supplied it is also possible to filter by time index or date, depending on the format of the 'Time' variable.
Defaults to \code{FALSE}.}

\item{lagPlot, }{Plots the dependent variable (y-axis) against a lagged version of itself (x-axis). The lag stands for the amount of observations in between the dependent variable and the lagged version of itself. The regression line is the autoregression at the specified lag.
Defaults to \code{FALSE}.}

\item{pacf, }{Plots the partial autocorrelation for a specified number of lags.
Defaults to \code{FALSE}.}

\item{time, }{Optional. Can either be an ordinal variable indicating the order of the observations, or a text variable indicating the date/time stamp of the observations. Combined date and time values should be in the standard format 'YYYY-MM-DD HH:MM:SS', where seconds (':SS') can also be omitted. Date-only values should be in the format 'YYYY-MM-DD'. If a time variable is not supplied, the row order of the data is used.}

\item{timeSeriesPlot, }{Plots the dependent variable (y-axis) over time (x-axis).
Defaults to \code{TRUE}.}
}
\description{
Time Series Descriptives allows the user to obtain descriptive statistics and descriptive plots for univariate time-series.
}
